+++ example/lib/examples/customizable_example/customizable_example.dart
@@ -23,7 +23,7 @@ class CustomizableExample extends StatelessWidget {
               shadowColor: Colors.grey[100]?.withOpacity(0.1),
               backgroundColor: Colors.white,
               title: Text(
-                'GroupButton 4.8.0',
+                'GroupButton 5.0.0',
                 style: Theme.of(context).textTheme.headline6,
               ),
             ),
@@ -34,7 +34,8 @@ class CustomizableExample extends StatelessWidget {
                 buttons: customizableController.buttons
                     .map((i) => '${i + 1}')
                     .toList(),
-                onSelected: (i, selected) => debugPrint('Button #$i $selected'),
+                onSelected: (val, i, selected) =>
+                    debugPrint('Button: $val index: $i $selected'),
               ),
             ),
             bottomNavigationBar: GroupButtonBottomPanel(